// src/models/contato.model.ts
export interface Contato {
    id: number;
    nome : string;
    isFavorito: boolean;
    email: string;
    cpf: string;
    cep: string;
    numero: number;
    logradouro: string;
    complemento: string;
    bairro: string;
    localidade : string;
    uf: string;
    gia: string;
    ddd: string;
    siafi: string;
  }